Undo weird change of renaming ValueTransformerReplaceIri, now with both replace all and first functionality
